:recycle: Move some static functionality of the menu preview to its own file
